movie_model_code/requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
numpy==1.26.4
pandas==2.2.1
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tzdata==2024.1
Werkzeug==3.0.1
sqlalchemy==2.0.28
python-dotenv==1.0.1
dnspython==2.6.1
email_validator==2.1.1
requests==2.31.0

# Flask Extensions
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
flask-paginate==2024.4.12

# Pytorch
torch==2.2.1
scikit-learn==1.4.1.post1
pytorch-lightning==2.2.1


plotly==5.21.0